requirements.txt
fastapi[all]
uvicorn[standard]
pydantic[email]
pydantic-settings
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
alembic
python-dotenv
python-multipart
jinja2
passlib[bcrypt]
python-jose[cryptography]
python-dateutil
aiosqlite
celery
redis
prometheus-fastapi-instrumentator
prometheus-client
requests
fpdf2
keyring
slowapi
